Fixed voxel raycasting system
This commit and the previous has fixed the issues of block placement and
deletion. The raycasting algorithm has been fixed to be used comfortably,
and the block placement and deletion algorithms have been configured to
work across chunks and in the current chunk well.

Nice commit, though I don't have time to improve other things. I have thought
of a relay system for controls that buffers key state every frame, and
determines when a key press can be registered based on its previous state.
If the previous frame was pressed, this frame is ignored. Otherwise, the
key press is registered. I'm pretty sure this is the way that most games work,
and this solution doesn't rely on setting limits of operation per frame.
I did benefit from setting a limit though, at least to make sure that block
presses aren't a problem.

Will look further into code standards, documentation, etc.