Added graphics submodule, refurbished makefile system
Didn't realize it but half the makefiles use the same template so I
imprinted it as a standard in common.mk and just used the simple
makefile writing system in each one.  Big time save that is.
Including that, I will not be adding my dev-link to
bin/output/wawacraft-evolved because not all users might agree with me
on that.

Gotta write lots and lots of documentation, and revise existing
documentation. But at least I have the ability to control an OpenGL
context.

I have one more thing to add, and that is to also solve the circular
dependency temporary solution, by moving GLFWwindow*
windows::window::get_window_handle into a reserved function somewhere
else.  Should I move all of windows::reserve namespace or just make it
for one file?