Fixed dependency on glad to be local, and made makefile run properly
For a long enough time, the makefile system was making a bad mistake of
not putting LFLAGS and LPATHS ahead of the central argument, which
breaks ld in some cases. Now this has been fixed up.  Thanks to the
style of makefile, this isn't a big change.

I made this change after switching to debian, because installing my
project isn't exactly the easiest on a new linux distro it seems...